Update nptag taskflow
