+++ examples/text_to_knowledge/nptag/README.md
@@ -6,3 +6,22 @@
 ## 模型结构
 
 模型使用[ERNIE-CTM](../ernie-ctm)+Prompt方式对名词短语进行分类
+
+## 在论文中引用NPTag
+
+如果您的工作成果中使用了NPTag，请增加下述引用。我们非常乐于看到解语对您的工作带来帮助。
+
+```
+@article{zhao2020TermTree,
+    title={TermTree and Knowledge Annotation Framework for Chinese Language Understanding},
+    author={Zhao, Min and Qin, Huapeng and Zhang, Guoxin and Lyu, Yajuan and Zhu, Yong},
+    technical report={Baidu, Inc. TR:2020-KG-TermTree},
+    year={2020}
+}
+```
+
+
+
+## 问题与反馈
+
+解语在持续优化中，如果您有任何建议或问题，欢迎提交issue到Github。
